requirements.txt
PyQt5
pyserial
siot
Pillow
